Enable whitelisting of slime island world gen dimensions
This change allows the user to toggle whether the list of slime island
dimension identifiers represents a blacklist or a whitelist. This allows
users with a large number of (possibly generated) dimensions to limit
the dimensions in which slime island generation takes place to a smaller
number without manually enumerating each one.

The following changes to the mod configuration options have been made:

 - `slimeIslandBlacklist` has been deprecated
 - `slimeIslandDimensions` has been added
 - `slimeIslandDimensionsIsBlacklist` has been added

`slimeIslandBlacklist` behaves as it did before, but is now empty by
default.

`slimeIslandDimensions` describes the list of dimensions in which slime
island generation will be whitelisted/blacklisted and defaults to the
old default value of `slimeIslandBlacklist`, which was `-1` and `1` (the
Nether and the End).

`slimeIslandDimensionsIsBlacklist` describes whether
`slimeIslandDimensions` represents a blacklist or a whitelist, and
defaults to `true` (blacklist).

If `slimeIslandBlacklist` has been modified, as may be the case when
upgrading from an older version of the mod, both `slimeIslandDimensions`
and `slimeIslandDimensionIsBlacklist` are ignored for backwards
compatibility.